scripts/requirements.txt
boto3>=1.37.5
matplotlib>=3.7.0
python-dotenv>=1.0.1
